Adds label_color as an option to line charts.
This commit adds a setting for the label colour in line charts, allowing users to override the default black label with a hex colour when creating a line chart. This is done by overriding the LabelSet's "Draw" method for Y axis labels, and by adding an @option to the "draw_x_label" LineChart class.

If no label colour is provided the default colour '#333' is used.