Adds axis names to line charts
This commit adds axis names to line charts, allowing you to show labels with
the names underneath the graph. You can name graph axis by doing the following:

	chart = new Charts.LineChart('graph', {
		y_axis_name:  'Conversion Rate',
		x_axis_name:  'Date',
		axis_name_color: '#aaa',
		axis_name_size:  10,
		x_padding: 60,
	});

There are options for the axis names, the axis font colour and the axis text
size to differentiate these from the labels for data values.

Note that right now you'll ahve to tweak X padding to work with your text size
for the Y axis.