Adds legend to line graph w/ hover & toggle
This commit adds a non customisable legend to the graph via a 'line_name'
option for each line. It takes the colour of each line and the optional
'line_name' argument and creates a legend in the top right corner of the
graph showing what each line represents.

When you hover over either a thumbnail or label for a line, the line on the
graph glows (with a drop-shadow), and you can toggle a line's visibility by
clicking its label/thumbnail in the legend.

This adds the following methods to the Line class:

* glow - adds a glow to the line and dots
* remove_glow - removes the glow from the line and dots
* hide - hides the line, dots and area
* show - shows the line, dots and area
* toggle - toggles the line, dot and area visibility

These methods provide the basis for our extra functionality and can be called on
the line directly.

There's also a @lines property of the LineChart class which saves all Lines
created when drawing the graph.